remove navbar utils
