+++ frontend/pages/buttons/FriendsButton.js
@@ -1,5 +1,4 @@
-import { Component } from "../Component.js";
-import { NavbarUtils } from "../utils/NavbarUtils.js";
+import { Component } from "../../Component.js";
 
 export class FriendsButton extends Component {
 	constructor() {
@@ -43,7 +42,6 @@ export class FriendsButton extends Component {
 		const sidebar = document.querySelector("friends-sidebar-component");
 		if (sidebar) {
 			sidebar.toggleVisibility();
-			NavbarUtils.hideCollapse();
 		}
 	}
 }
