+++ frontend/pages/buttons/ThemeButton.js
@@ -1,6 +1,5 @@
 import { Component } from "../Component.js";
 import { Theme } from "../components";
-import { NavbarUtils } from "../utils/NavbarUtils.js";
 
 export class ThemeButton extends Component {
 	constructor() {
@@ -54,6 +53,5 @@ export class ThemeButton extends Component {
 		}
 		this.icon.classList.toggle("bi-moon-fill");
 		this.icon.classList.toggle("bi-sun-fill");
-		NavbarUtils.hideCollapse();
 	}
 }
