--- frontend/pages/buttons/index.js
@@ -1,11 +1,0 @@
-import {FriendsButton} from './FriendsButton.js';
-import {IntraButton} from '../auth/IntraButton.js';
-import {MultiplayerButton} from './MultiplayerButton.js';
-import {ThemeButton} from './ThemeButton.js';
-
-export {
-  FriendsButton,
-  IntraButton,
-  MultiplayerButton,
-  ThemeButton,
-};
